requirements.txt
ipykernel
Flask==3.1.0
joblib==1.4.2
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
pymongo==4.11
Requests==2.32.3
scikit_learn==1.6.1
tensorflow-cpu
